+++ footpedalkeyboard-driver/CHANGELOG.md
@@ -2,5 +2,10 @@
 
 ## Versions
 
+1.1.0, 2020-10-31
+- Added support for process name based active application matching
+- Fixed key mappings for Spotify to work on process name 
+- Code refactor and cleanup
+
 1.0.0, 2020-10-27
 - Added key mappings support for Winamp